"interpolate_zeros_nans" in commons rather than in synchro and triang
