+++ Pose2Sim/triangulation.py
@@ -44,15 +44,14 @@ import pandas as pd
 import cv2
 import toml
 from tqdm import tqdm
-from scipy import interpolate
 from collections import Counter
 from anytree import RenderTree
 from anytree.importer import DictImporter
 import logging
 
 from Pose2Sim.common import retrieve_calib_params, computeP, weighted_triangulation, \
-    reprojection, euclidean_distance, sort_people_sports2d, \
-    sort_stringlist_by_last_number, min_with_single_indices, zup2yup, convert_to_c3d
+    reprojection, euclidean_distance, sort_people_sports2d, interpolate_zeros_nans, \
+    sort_stringlist_by_last_number, zup2yup, convert_to_c3d
 from Pose2Sim.skeletons import *
 
 
@@ -68,52 +67,6 @@ __status__ = "Development"
 
 
 ## FUNCTIONS
-def interpolate_zeros_nans(col, *args):
-    '''
-    Interpolate missing points (of value zero),
-    unless more than N contiguous values are missing.
-
-    INPUTS:
-    - col: pandas column of coordinates
-    - args[0] = N: max number of contiguous bad values, above which they won't be interpolated
-    - args[1] = kind: 'linear', 'slinear', 'quadratic', 'cubic'. Default: 'cubic'
-
-    OUTPUT:
-    - col_interp: interpolated pandas column
-    '''
-
-    if len(args)==2:
-        N, kind = args
-    if len(args)==1:
-        N = np.inf
-        kind = args[0]
-    if not args:
-        N = np.inf
-    
-    # Interpolate nans
-    mask = ~(np.isnan(col) | col.eq(0)) # true where nans or zeros
-    idx_good = np.where(mask)[0]
-    if len(idx_good) <= 4:
-        return col
-        
-    if 'kind' not in locals(): # 'linear', 'slinear', 'quadratic', 'cubic'
-        f_interp = interpolate.interp1d(idx_good, col[idx_good], kind="linear", bounds_error=False)
-    else:
-        f_interp = interpolate.interp1d(idx_good, col[idx_good], kind=kind, fill_value='extrapolate', bounds_error=False)
-    col_interp = np.where(mask, col, f_interp(col.index)) #replace at false index with interpolated values
-    
-    # Reintroduce nans if length of sequence > N
-    idx_notgood = np.where(~mask)[0]
-    gaps = np.where(np.diff(idx_notgood) > 1)[0] + 1 # where the indices of true are not contiguous
-    sequences = np.split(idx_notgood, gaps)
-    if sequences[0].size>0:
-        for seq in sequences:
-            if len(seq) > N: # values to exclude from interpolation are set to false when they are too long 
-                col_interp[seq] = np.nan
-    
-    return col_interp
-
-
 def count_persons_in_json(file_path):
     with open(file_path, 'r') as file:
         data = json.load(file)
@@ -880,7 +833,7 @@ def triangulate_all(config_dict):
     if interpolation_kind != 'none':
         for n in range(nb_persons_to_detect):
             try:
-                Q_tot[n] = Q_tot[n].apply(interpolate_zeros_nans, axis=0, args = [interp_gap_smaller_than, interpolation_kind])
+                Q_tot[n] = Q_tot[n].apply(interpolate_zeros_nans, axis=0, args=[interp_gap_smaller_than, interpolation_kind])
             except:
                 logging.info(f'Interpolation was not possible for person {n}. This means that not enough points are available, which is often due to a bad calibration.')
     # Fill non-interpolated values with last valid one
